perf(nvim): remap keybinding for formatting file

+++ .config/nvim/lua/jeraldlyh/plugins/lsp/lspconfig.lua
@@ -118,7 +118,7 @@ return {
                 },
               },
               format = {
-                ["quote-style"] = "double",
+                ["quote-style"] = "single",
               },
             },
           },
